Test tip building
