+++ packages/backend/src/api/routers/ActivityRouter.ts
@@ -1,6 +1,6 @@
 import Router from '@koa/router'
 
-import { ActivityController } from '../controllers/ActivityController'
+import { ActivityController } from '../controllers/activity/ActivityController'
 
 export function createActivityRouter(activityController: ActivityController) {
   const router = new Router()
