+++ packages/backend/test/api/controllers/activity/ActivityController.test.ts
@@ -1,0 +1,137 @@
+import { mock } from '@l2beat/common'
+import {
+  ActivityApiChartPoint,
+  ActivityApiResponse,
+  ProjectId,
+  UnixTime,
+} from '@l2beat/types'
+import { expect } from 'earljs'
+
+import { ActivityController } from '../../../../src/api/controllers/activity/ActivityController'
+import { RpcTransactionUpdater } from '../../../../src/core/transaction-count/RpcTransactionUpdater'
+import { StarkexTransactionUpdater } from '../../../../src/core/transaction-count/StarkexTransactionUpdater'
+import { TransactionCounter } from '../../../../src/core/transaction-count/TransactionCounter'
+import { ZksyncTransactionUpdater } from '../../../../src/core/transaction-count/ZksyncTransactionUpdater'
+
+describe(ActivityController.name, () => {
+  describe(ActivityController.prototype.getTransactionActivity.name, () => {
+    it('returns empty if no projects track activity', async () => {
+      const controller = new ActivityController(
+        [],
+        mock<TransactionCounter>({
+          async getFullySyncedDailyCounts() {
+            return []
+          },
+        }),
+      )
+      expect(await controller.getTransactionActivity()).toEqual(
+        formatActivity({ combined: [], projects: {}, ethereum: [] }),
+      )
+    })
+
+    it('groups data', async () => {
+      const controller = new ActivityController(
+        [
+          mock<RpcTransactionUpdater>({
+            projectId: ProjectId('optimism'),
+            async getFullySyncedDailyCounts() {
+              return [
+                { count: 1, timestamp: new UnixTime(0).add(1, 'days') },
+                { count: 2, timestamp: new UnixTime(0).add(2, 'days') },
+                { count: 3, timestamp: new UnixTime(0).add(3, 'days') },
+              ]
+            },
+          }),
+          mock<StarkexTransactionUpdater>({
+            projectId: ProjectId('dydx'),
+            async getFullySyncedDailyCounts() {
+              return [
+                { count: 4, timestamp: new UnixTime(0) },
+                { count: 5, timestamp: new UnixTime(0).add(1, 'days') },
+                { count: 6, timestamp: new UnixTime(0).add(2, 'days') },
+              ]
+            },
+          }),
+          mock<ZksyncTransactionUpdater>({
+            projectId: ProjectId('zksync'),
+            async getFullySyncedDailyCounts() {
+              return [
+                { count: 7, timestamp: new UnixTime(0).add(2, 'days') },
+                { count: 8, timestamp: new UnixTime(0).add(3, 'days') },
+                { count: 9, timestamp: new UnixTime(0).add(4, 'days') },
+              ]
+            },
+          }),
+        ],
+        mock<TransactionCounter>({
+          async getFullySyncedDailyCounts() {
+            return [
+              { count: 100, timestamp: new UnixTime(0) },
+              { count: 200, timestamp: new UnixTime(0).add(1, 'days') },
+              { count: 300, timestamp: new UnixTime(0).add(2, 'days') },
+              { count: 400, timestamp: new UnixTime(0).add(3, 'days') },
+              { count: 500, timestamp: new UnixTime(0).add(4, 'days') },
+            ]
+          },
+        }),
+      )
+
+      expect(await controller.getTransactionActivity()).toEqual(
+        formatActivity({
+          combined: [
+            [new UnixTime(0), 4],
+            [new UnixTime(0).add(1, 'days'), 6],
+            [new UnixTime(0).add(2, 'days'), 15],
+          ],
+          projects: {
+            dydx: [
+              [new UnixTime(0), 4],
+              [new UnixTime(0).add(1, 'days'), 5],
+              [new UnixTime(0).add(2, 'days'), 6],
+            ],
+            optimism: [
+              [new UnixTime(0).add(1, 'days'), 1],
+              [new UnixTime(0).add(2, 'days'), 2],
+            ],
+            zksync: [[new UnixTime(0).add(2, 'days'), 7]],
+          },
+          ethereum: [
+            [new UnixTime(0), 100],
+            [new UnixTime(0).add(1, 'days'), 200],
+            [new UnixTime(0).add(2, 'days'), 300],
+          ],
+        }),
+      )
+    })
+  })
+})
+
+function formatActivity({
+  combined,
+  projects,
+  ethereum,
+}: {
+  combined: ActivityApiChartPoint[]
+  projects: Record<string, ActivityApiChartPoint[]>
+  ethereum: ActivityApiChartPoint[]
+}): ActivityApiResponse {
+  return {
+    combined: {
+      types: ['timestamp', 'daily tx count'],
+      data: combined,
+    },
+    ethereum: {
+      types: ['timestamp', 'daily tx count'],
+      data: ethereum,
+    },
+    projects: Object.entries(projects).reduce((acc, cur) => {
+      return {
+        ...acc,
+        [cur[0]]: {
+          types: ['timestamp', 'daily tx count'],
+          data: cur[1],
+        },
+      }
+    }, {}),
+  }
+}
